Make clearing by type actually work
Clearing by type in SurfaceFragment didn't actually work. Instead,
it followed a code path intended for setting and created an annotation
of that type with no data, then tried to clear that. What we really
want to do there is clear anything with that type.

This fixes the bug where unbolding of text that was already bold in
the article didn't work.

Bug: 47680
Change-Id: I77f00e63c8732420063b0453fede7f453083c913